popovers: Refactor render_user_info_popover function.
This commit separates the render_user_info_popover function from
popovers.js and moves it to popover_menus_data.js. This function now
returns an args object containing all parameters for the user card.
This object will be passed to `render_user_info_popover_content` in
upcoming commits. Shifted some related functions alongside.

Removed click handlers for user card options which will be added
to tippy user card's onMount function in next commits.

Removed popover.test.js since there was only one test that was for
popover but now user card is converted to tippy it won't work now.
Modified people.test.js to add a case as after removing
popover.test.js people.js was losing 100% coverage.

Fixes #24280
