+++ web/src/popover_menus_data.js
@@ -1,8 +1,14 @@
 /* This module provides relevant data to render popovers that require multiple args.
    This helps keep the popovers code small and keep it focused on rendering side of things. */
 
+import ClipboardJS from "clipboard";
+import {parseISO} from "date-fns";
+import $ from "jquery";
+import tippy from "tippy.js";
+
 import * as resolved_topic from "../shared/src/resolved_topic";
 
+import * as buddy_data from "./buddy_data";
 import * as hash_util from "./hash_util";
 import {$t} from "./i18n";
 import * as message_edit from "./message_edit";
@@ -10,10 +16,16 @@ import * as message_lists from "./message_lists";
 import * as muted_users from "./muted_users";
 import * as narrow_state from "./narrow_state";
 import {page_params} from "./page_params";
+import * as people from "./people";
+import * as settings_config from "./settings_config";
 import * as settings_data from "./settings_data";
 import * as starred_messages from "./starred_messages";
 import * as stream_data from "./stream_data";
 import * as sub_store from "./sub_store";
+import * as timerender from "./timerender";
+import * as user_profile from "./user_profile";
+import {user_settings} from "./user_settings";
+import * as user_status from "./user_status";
 import * as user_topics from "./user_topics";
 
 export function get_actions_popover_content_context(message_id) {
@@ -145,4 +157,179 @@ export function get_topic_popover_content_context({stream_id, topic_name, url}) 
         has_starred_messages,
         url,
     };
+}
+
+export function clipboard_enable(arg) {
+    // arg is a selector or element
+    // We extract this function for testing purpose.
+    if (arg) {
+        return new ClipboardJS(arg);
+    }
+    return null;
+}
+
+function copy_email_handler(e) {
+    const $email_el = $(e.trigger.parentElement);
+    const $copy_icon = $email_el.find("i");
+
+    // only change the parent element's text back to email
+    // and not overwrite the tooltip.
+    const email_textnode = $email_el[0].childNodes[2];
+
+    $email_el.addClass("email_copied");
+    email_textnode.nodeValue = $t({defaultMessage: "Email copied"});
+
+    setTimeout(() => {
+        $email_el.removeClass("email_copied");
+        email_textnode.nodeValue = $copy_icon.attr("data-clipboard-text");
+    }, 1500);
+    e.clearSelection();
+}
+
+export function init_email_clipboard() {
+    /*
+        This shows (and enables) the copy-text icon for folks
+        who have names that would overflow past the right
+        edge of our user mention popup.
+    */
+    $(".user_popover_email").each(function () {
+        if (this.clientWidth < this.scrollWidth) {
+            const $email_el = $(this);
+            const $copy_email_icon = $email_el.find("i");
+
+            /*
+                For deactivated users, the copy-email icon will
+                not even be present in the HTML, so we don't do
+                anything.  We don't reveal emails for deactivated
+                users.
+            */
+            if ($copy_email_icon[0]) {
+                $copy_email_icon.removeClass("hide_copy_icon");
+                const copy_email_clipboard = clipboard_enable($copy_email_icon[0]);
+                copy_email_clipboard.on("success", copy_email_handler);
+            }
+        }
+    });
+}
+
+export function init_email_tooltip(user) {
+    /*
+        This displays the email tooltip for folks
+        who have names that would overflow past the right
+        edge of our user mention popup.
+    */
+
+    $(".user_popover_email").each(function () {
+        if (this.clientWidth < this.scrollWidth) {
+            tippy(this, {
+                placement: "bottom",
+                content: people.get_visible_email(user),
+                interactive: true,
+            });
+        }
+    });
+}
+
+export function load_medium_avatar(user, $elt) {
+    const user_avatar_url = people.medium_avatar_url_for_person(user);
+    const sender_avatar_medium = new Image();
+
+    sender_avatar_medium.src = user_avatar_url;
+    $(sender_avatar_medium).on("load", function () {
+        $elt.css("background-image", "url(" + $(this).attr("src") + ")");
+    });
+}
+
+export function render_user_info_popover(
+    user,
+    is_sender_popover,
+    has_message_context,
+    template_class,
+) {
+    const is_me = people.is_my_user_id(user.user_id);
+
+    let invisible_mode = false;
+
+    if (is_me) {
+        invisible_mode = !user_settings.presence_enabled;
+    }
+
+    const muting_allowed = !is_me && !user.is_bot;
+    const is_active = people.is_active_user_for_popover(user.user_id);
+    const is_system_bot = user.is_system_bot;
+    const status_text = user_status.get_status_text(user.user_id);
+    const status_emoji_info = user_status.get_status_emoji(user.user_id);
+    const spectator_view = page_params.is_spectator;
+
+    // TODO: The show_manage_menu calculation can get a lot simpler
+    // if/when we allow muting bot users.
+    const can_manage_user = page_params.is_admin && !is_me && !is_system_bot;
+    const show_manage_menu = !spectator_view && (muting_allowed || can_manage_user);
+
+    let date_joined;
+    if (spectator_view) {
+        date_joined = timerender.get_localized_date_or_time_for_format(
+            parseISO(user.date_joined),
+            "dayofyear_year",
+        );
+    }
+    // Filtering out only those profile fields that can be display in the popover and are not empty.
+    const field_types = page_params.custom_profile_field_types;
+    const display_profile_fields = page_params.custom_profile_fields
+        .map((f) => user_profile.get_custom_profile_field_data(user, f, field_types))
+        .filter((f) => f.display_in_profile_summary && f.value !== undefined && f.value !== null);
+
+    const args = {
+        class: template_class,
+        user_is_guest: user.is_guest,
+        user_avatar: people.small_avatar_url_for_person(user),
+        invisible_mode,
+        can_send_private_message:
+            is_active &&
+            !is_me &&
+            page_params.realm_private_message_policy !==
+                settings_config.private_message_policy_values.disabled.code,
+        display_profile_fields,
+        has_message_context,
+        is_active,
+        is_bot: user.is_bot,
+        is_me,
+        is_sender_popover,
+        pm_with_url: hash_util.pm_with_url(user.email),
+        user_circle_class: buddy_data.get_user_circle_class(user.user_id),
+        private_message_class: "compose_private_message",
+        sent_by_uri: hash_util.by_sender_url(user.email),
+        show_manage_menu,
+        user_email: user.delivery_email,
+        user_full_name: user.full_name,
+        user_id: user.user_id,
+        user_last_seen_time_status: buddy_data.user_last_seen_time_status(user.user_id),
+        user_time: people.get_user_time(user.user_id),
+        user_type: people.get_user_type(user.user_id),
+        status_content_available: Boolean(status_text || status_emoji_info),
+        status_text,
+        status_emoji_info,
+        user_mention_syntax: people.get_mention_syntax(user.full_name, user.user_id),
+        date_joined,
+        spectator_view,
+    };
+
+    if (user.is_bot) {
+        const bot_owner_id = user.bot_owner_id;
+        if (is_system_bot) {
+            args.is_system_bot = is_system_bot;
+        } else if (bot_owner_id) {
+            const bot_owner = people.get_by_user_id(bot_owner_id);
+            args.bot_owner = bot_owner;
+        }
+    }
+
+    // Note: We pass the normal-size avatar in initial rendering, and
+    // then query the server to replace it with the medium-size
+    // avatar.  The purpose of this double-fetch approach is to take
+    // advantage of the fact that the browser should already have the
+    // low-resolution image cached and thus display a low-resolution
+    // avatar rather than a blank area during the network delay for
+    // fetching the medium-size one.
+    return args;
 }
